Merge in NEXTOCEAN/vcw4next from feature/NEXO-318-fix-criteria-update-on-criteria-with-negative-source to develop

* commit 'da53f523efaa6745bd71b84ca9ee85bfb3e46e6a':
  fix: added validation when updating a negative source with another negative source

+++ packages/backend/src/main/java/pt/com/deimos/vcwapi/service/IdeaAndCriteriaService.java
@@ -104,23 +104,23 @@ public class IdeaAndCriteriaService {
     /*SourceEntity oldSource = oldIdeaAndCriteria.getSource();
     SourceDTO newSourceInfo = editedInfo.getSource();
     SourceEntity newSource;
-    if (newSourceInfo == null) {
-      oldIdeaAndCriteria.setSource(null);
-      oldSource.removeIdeaAndCriteria(oldIdeaAndCriteria);
+    if (newSourceInfo != null && oldSource != null){
+      BeanUtils.copyProperties(newSourceInfo, oldSource);
+      oldSource.setUpdatedAt(LocalDateTime.now());
+      oldSource.setUpdatedBy(userId);
+    }
+    if (newSourceInfo == null && oldSource != null) {
+      oldCriteria.setSource(null);
+      oldSource.removeCriteria(oldCriteria);
       this.sourceRepository.delete(oldSource);
     }
-    else if (oldSource == null) {
+    else if (oldSource == null && newSourceInfo != null) {
       newSource = new SourceEntity();
       BeanUtils.copyProperties(newSourceInfo, newSource);
       newSource.setCreatedBy(userId);
       newSource.setUpdatedAt(LocalDateTime.now());
       newSource.setUpdatedBy(userId);
-      oldIdeaAndCriteria.setSource(newSource);
-    }
-    else {
-      BeanUtils.copyProperties(newSourceInfo, oldSource);
-      oldSource.setUpdatedAt(LocalDateTime.now());
-      oldSource.setUpdatedBy(userId);
+      oldCriteria.setSource(newSource);
     }*/
 
     Optional<CriteriaEntity> c = this.criteriaRepository.findById(editedInfo.getCriteriaId());
